requirements.txt
tushare>=1.2.89
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
colorama>=0.4.6

